cleanup old unit tests
