--- core/src/test/java/adapting/CustomType.java
@@ -1,4 +1,0 @@
-package adapting;
-
-public class CustomType {
-}
