--- core/src/test/java/commands/UnsupportedType.java
@@ -1,4 +1,0 @@
-package commands;
-
-public class UnsupportedType {
-}
