+++ gradle/libs.versions.toml
@@ -5,19 +5,16 @@
 com-google-code-gson-gson = "2.10.1"
 net-dv8tion-jda = "5.5.1"
 org-jetbrains-annotations = "24.0.1"
-org-junit-jupiter-junit-jupiter-api = "5.9.2"
-org-junit-jupiter-junit-jupiter-engine = "5.9.2"
 org-reflections-reflections = "0.10.2"
 org-slf4j-slf4j-api = "2.0.5"
 com-google-inject-guice = "7.0.0"
-
+org-junit-jupiter = "5.13.0"
 
 [libraries]
 com-google-code-gson-gson = { module = "com.google.code.gson:gson", version.ref = "com-google-code-gson-gson" }
 net-dv8tion-jda = { module = "net.dv8tion:JDA", version.ref = "net-dv8tion-jda" }
 org-jetbrains-annotations = { module = "org.jetbrains:annotations", version.ref = "org-jetbrains-annotations" }
-org-junit-jupiter-junit-jupiter-api = { module = "org.junit.jupiter:junit-jupiter-api", version.ref = "org-junit-jupiter-junit-jupiter-api" }
-org-junit-jupiter-junit-jupiter-engine = { module = "org.junit.jupiter:junit-jupiter-engine", version.ref = "org-junit-jupiter-junit-jupiter-engine" }
+org-junit-jupiter = { module = "org.junit.jupiter:junit-jupiter", version.ref = "org-junit-jupiter" }
 org-reflections-reflections = { module = "org.reflections:reflections", version.ref = "org-reflections-reflections" }
 org-slf4j-slf4j-api = { module = "org.slf4j:slf4j-api", version.ref = "org-slf4j-slf4j-api" }
 com-google-inject-guice = { module = "com.google.inject:guice", version.ref = "com-google-inject-guice" }
